requirements.txt

python-twitter>=1.9.0
django>=1.4
django-delegate>=0.2.2
django-docfield-couchdb>=0.1.1
django-appconf>=0.4
django-tagging>=0.3.0
